+++ src/moves_generator.rs
@@ -48,10 +48,7 @@ impl MovesGenerator for Game {
         let piece = self.board[m.from() as usize];
         let capture = self.board[m.to() as usize]; // TODO: En passant
 
-        debug_assert_eq!(piece.color(), side);
-
         self.bitboards[piece as usize].toggle(m.from());
-        debug_assert!(!self.bitboards[piece as usize].get(m.from()));
         self.board[m.from() as usize] = EMPTY;
         new_position.hash ^= self.zobrist.positions[piece as usize][m.from() as usize];
         new_position.halfmoves_count += 1;
@@ -145,9 +142,7 @@ impl MovesGenerator for Game {
         }
 
         self.bitboards[side as usize].toggle(m.from());
-        debug_assert!(!self.bitboards[side as usize].get(m.from()));
         self.bitboards[side as usize].toggle(m.to());
-        debug_assert!(self.bitboards[side as usize].get(m.to()));
 
         // if m.is_capture() {
         if capture != EMPTY {